# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WeedTown API",
    "version": "2.4.0",
    "description": "API de WeedTown, red social de la comunidad cannábica de México. Identidad federada con Mastodon (OAuth 2.0) y sesión mediante JWT propio en el header Authorization. Seguridad: headers via helmet, CORS restringido al frontend, rate limiting (300 peticiones/15 min general; 20/15 min en el flujo OAuth), payload JSON limitado a 100 kB e imágenes a 5 MB. El perfil público (GET /api/profile/{id}) no expone PII (email, teléfono, nombre real, edad, nacimiento ni género). Las rutas de mercado y admin son stubs pendientes de implementación."
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT emitido por el backend tras el flujo OAuth de Mastodon (expira en 7 días)."
      }
    },
    "schemas": {
...
      "Post": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "authorId": {
            "type": "integer"
          },
          "content": {
            "type": "string"
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "author": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              },
              "name": {
                "type": "string"
              },
              "avatar": {
                "type": "string",
                "nullable": true
              }
            }
          },
          "hashtags": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "postId": {
                  "type": "integer"
                },
                "hashtagId": {
                  "type": "integer"
                },
                "hashtag": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "tag": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "reactions": {
            "$ref": "#/components/schemas/ReactionCounts"
          },
          "myReaction": {
            "$ref": "#/components/schemas/ReactionType"
          },
          "commentCount": {
            "type": "integer"
          }
        }
      },
      "ReactionType": {
        "type": "string",
        "nullable": true,
        "enum": [
          "LIKE",
          "ROLA",
          "INTERESA",
          "MOLESTA"
        ],
        "description": "LIKE = Me gusta, ROLA = Me rola, INTERESA = Me interesa, MOLESTA = Me molesta"
      },
      "ReactionCounts": {
        "type": "object",
        "properties": {
          "LIKE": {
            "type": "integer"
          },
          "ROLA": {
            "type": "integer"
          },
          "INTERESA": {
            "type": "integer"
          },
          "MOLESTA": {
            "type": "integer"
          }
        }
      },
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "postId": {
            "type": "integer"
          },
          "content": {
            "type": "string"
          },
          "image": {
            "type": "string",
            "nullable": true,
            "description": "URL de imagen adjunta (opcional)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "author": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              },
              "name": {
                "type": "string"
              },
              "avatar": {
                "type": "string",
                "nullable": true
              }
            }
          },
          "reactions": {
            "$ref": "#/components/schemas/ReactionCounts"
          },
          "myReaction": {
            "$ref": "#/components/schemas/ReactionType"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    }
  },
...
  "paths": {
...
    "/api/posts": {
      "get": {
        "tags": [
          "Posts"
        ],
        "summary": "Feed paginado de posteos (20 por página)",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Página del feed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "posts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Post"
                      }
                    },
                    "page": {
                      "type": "integer"
                    },
                    "totalPages": {
                      "type": "integer"
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Posts"
        ],
        "summary": "Crear posteo (el autor sale del token)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string",
                    "description": "URL de imagen (opcional)"
                  },
                  "hashtags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "cultura",
                      "420",
                      "cdmx"
                    ]
                  }
                },
                "required": [
                  "content"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Posteo creado con autor y hashtags",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "400": {
            "description": "Faltan campos requeridos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/search": {
      "get": {
        "tags": [
          "Posts"
        ],
        "summary": "Buscar posteos por contenido o nombre de autor",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resultados de búsqueda",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Post"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/posts/{id}/reaction": {
      "post": {
        "tags": [
          "Reacciones"
        ],
        "summary": "Reaccionar a un posteo (misma reacción = quitar, distinta = reemplazar)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "$ref": "#/components/schemas/ReactionType"
                  }
                },
                "required": [
                  "type"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reacción registrada; devuelve myReaction y conteos frescos"
          },
          "400": {
            "description": "Tipo de reacción inválido"
          },
          "401": {
            "description": "No autenticado"
          },
          "404": {
            "description": "Post no encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "Reacciones"
        ],
        "summary": "Quitar la reacción propia de un posteo",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reacción eliminada; devuelve conteos frescos"
          },
          "401": {
            "description": "No autenticado"
          }
        }
      }
    },
    "/api/posts/{id}/like": {
      "post": {
        "tags": [
          "Reacciones"
        ],
        "summary": "Alias de compatibilidad: equivale a reaccionar con LIKE",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reacción LIKE registrada (o quitada si ya existía)"
          },
          "401": {
            "description": "No autenticado"
          }
        }
      }
    },
    "/api/posts/{id}/comment": {
      "post": {
        "tags": [
          "Comentarios"
        ],
        "summary": "Comentar un posteo",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string"
                  }
                },
                "required": [
                  "content"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Comentario creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          },
          "400": {
            "description": "Comentario vacío"
          },
          "401": {
            "description": "No autenticado"
          },
          "404": {
            "description": "Post no encontrado"
          }
        }
      }
    },
    "/api/posts/{id}/comments": {
      "get": {
        "tags": [
          "Comentarios"
        ],
        "summary": "Listar comentarios de un posteo (con conteos de reacciones)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comentarios del post",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "comments": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Comment"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/forum/subforums/{slug}/posts": {
      "get": {
        "tags": [
          "Foros"
        ],
        "summary": "Posts del subforo con orden estilo Reddit",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "hot",
                "new",
                "top"
              ],
              "default": "hot"
            },
            "description": "hot = puntaje con decaimiento temporal"
          },
          {
            "name": "period",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "day",
                "week",
                "month",
                "all"
              ],
              "default": "all"
            },
            "description": "Solo aplica a top"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Posts con score, reacciones, myReaction y commentCount"
          }
        }
      },
      "post": {
        "tags": [
          "Foros"
        ],
        "summary": "Publicar en el subforo (imagen opcional vía /api/media/upload)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string"
                  }
                },
                "required": [
                  "title",
                  "content"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post creado; notifica a seguidores del subforo"
          }
        }
      }
    },
    "/api/forum/posts/{id}": {
      "get": {
        "tags": [
          "Foros"
        ],
        "summary": "Detalle de un post del foro",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post con score y reacciones"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Foros"
        ],
        "summary": "Editar post propio (403 si es ajeno)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post actualizado"
          },
          "403": {
            "description": "No es tu contenido"
          }
        }
      },
      "delete": {
        "tags": [
          "Foros"
        ],
        "summary": "Eliminar post propio con todo su hilo (403 si es ajeno)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Eliminado"
          },
          "403": {
            "description": "No es tu contenido"
          }
        }
      }
    },
    "/api/forum/posts/{id}/comments": {
      "get": {
        "tags": [
          "Foros"
        ],
        "summary": "Hilo completo del post (el árbol se arma en el cliente)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comentarios con depth (0-2), parent.author, reacciones y deleted"
          }
        }
      },
      "post": {
        "tags": [
          "Foros"
        ],
        "summary": "Comentar o responder (parentId opcional; >3 niveles se aplana)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string"
                  },
                  "parentId": {
                    "type": "integer"
                  }
                },
                "required": [
                  "content"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Comentario creado; notifica al autor del post o del padre"
          }
        }
      }
    },
    "/api/forum/posts/{id}/reaction": {
      "post": {
        "tags": [
          "Foros"
        ],
        "summary": "Reaccionar al post: puntúa (+1 LIKE/ROLA/INTERESA, -1 MOLESTA); misma = quitar",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ myReaction, reactions, score }"
          }
        }
      },
      "delete": {
        "tags": [
          "Foros"
        ],
        "summary": "Quitar la reacción propia (ajusta el score)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ myReaction: null, reactions, score }"
          }
        }
      }
    },
...
  }
}